+++ main/createMatrix.h
@@ -22,14 +22,14 @@ public:
     void createCustomMatrix();
     void printMatrix(); // Prints the matrix
     void getSolutionsVector(); // the last integer of each row
-    int** getMatrix();
+    float** getMatrix();
     int getRows();
     int getColumns();
 
 private:
     int rows; // Number of rows in the matrix
     int columns; // Number of columns in the matrix
-    int** matrix; // Pointer to a 2D array (matrix)
+    float** matrix; // Pointer to a 2D array (matrix)
 
     void deallocateMatrix(); // Helper function to deallocate memory
     void checkIfNull();
